requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
distlib==0.3.1
Django==3.0.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
filelock==3.0.12
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.31
wrapt==1.12.1
